update timer tracker api

+++ pages/api/timer/category/delete.js
@@ -1,6 +1,9 @@
 import { prisma } from "@/auth";
 
 export default async function handler(req, res) {
+  if (req.method != "POST") {
+    return res.status(403).json({ message: "Request forbidden" });
+  }
   const { deleteList } = req.body;
   const result = await prisma.category.deleteMany({
     where: {
